requirements.txt
numpy
psutil
python-dotenv
pyyaml
requests
sqlalchemy
tqdm

